edit  'double' instead of 'int' in sub

+++ java_backend/j_calculator/src/main/java/com/example/j_calculator/CalculatorController.java
@@ -12,7 +12,7 @@ public class CalculatorController {
     }
 
     @GetMapping("/sub")
-    public String sub(@RequestParam int a, @RequestParam int b) {
+    public String sub(@RequestParam double a, @RequestParam double b) {
         return String.valueOf(a - b);
     }
